backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.38",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "uvicorn>=0.44.0",
]
